Added login required decorator to secure views

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -98,8 +99,13 @@ def product_details_page(request, product_id):
     return render(request, 'products/book_details_page.html', context)
 
 
+@login_required
 def add_product(request):
     """Add product to the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorr, but you do not have permissions to acces this page.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -120,7 +126,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
+    """Allows superuser to edit products ont he site"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorr, but you do not have permissions to access this page.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
 
     if request.method == 'POST':
@@ -144,8 +156,12 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """Deletes a product from the site"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorr, but you do not have permissions to acces this page.')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
